# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Praça.ai marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: registerConsumer
      tags: [profile]
      summary: Register a new consumer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerRegisterInput"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerSession"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginConsumer
      tags: [profile]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerLoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [profile]
      summary: Get current logged-in consumer
      responses:
        "200":
          description: Current consumer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /favorites/{productId}:
    post:
      operationId: toggleFavorite
      tags: [favorites]
      summary: Toggle product as favorite
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteToggleResult"
...
components:
  schemas:
    ConsumerRegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        password:
          type: string
      required:
        - name
        - email
        - password
    ConsumerLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    ConsumerSession:
      type: object
      properties:
        id:
          type: ["integer", "string"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discountPct:
          type: ["integer", "null"]
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        categorySlug:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        vendorLogoUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        salesCount:
          type: integer
        stock:
          type: integer
        isFavorited:
          type: boolean
        sizes:
          type: ["array", "null"]
          items:
            type: string
        deliveryDays:
          type: integer
        freeShipping:
          type: boolean
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - images
        - category
        - categorySlug
        - vendorId
        - vendorName
        - rating
        - reviewCount
        - salesCount
        - stock
        - isFavorited
        - deliveryDays
        - freeShipping
...
    CartItem:
      type: object
      properties:
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
      required:
        - productId
        - product
        - quantity
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - shipping
        - total
        - itemCount
    CartItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
      required:
        - productId
        - quantity
...
    FavoriteToggleResult:
      type: object
      properties:
        productId:
          type: string
        isFavorited:
          type: boolean
      required:
        - productId
        - isFavorited
...