# lib/api-zod/src/generated/types/consumerLoginInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Praça.ai marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumerLoginInput {
  email: string;
  password: string;
}