# lib/api-zod/src/generated/types/consumerSession.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Praça.ai marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumerSession {
  id: number | string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  createdAt?: string;
}